change midi-id's type to uint16_t and rename to connect-id
Other controllers use different prefix than 0x41.
Add ability to use controls with different prefix.

https://www.untergeek.de/2014/11/taming-arturias-beatstep-sysex-codes-for-programming-via-ipad/

+++ src/ks37-midi.h
@@ -69,8 +69,8 @@
 #define P_BANK4_KNOB4_MAX		0x6D  // 0-127
 #define P_BANK4_KNOB4_CHAN		0x6E  // 0-15, user=0x41
 
-int ks37_midi_read_control(snd_seq_t *seq, snd_seq_addr_t addr, uint8_t control_id, uint8_t *val);
-int ks37_midi_write_control(snd_seq_t *seq, snd_seq_addr_t addr, uint8_t control_id, uint8_t val);
+int ks37_midi_read_control(snd_seq_t *seq, snd_seq_addr_t addr, uint16_t control_id, uint8_t *val);
+int ks37_midi_write_control(snd_seq_t *seq, snd_seq_addr_t addr, uint16_t control_id, uint8_t val);
 int ks37_midi_connect(snd_seq_t *seq, snd_seq_addr_t addr);
 int ks37_midi_get_address(snd_seq_t *seq, const char* name, snd_seq_addr_t *addr);
 int ks37_midi_open(snd_seq_t **seq);
